Update BestPractices.md to remove alpine user
Since the `node` user was added to the alpine variant, there is no need to recommend that they need to create the `node` user. In fact, it would break. Only the `onbuild` variant does not have the user.

I also added the `USER node` syntax to show how to [set the user](https://docs.docker.com/engine/reference/builder/#user) in the dockerfile.